Move utils tests
